main/backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
pydantic>=2.0.0
pytesseract>=0.3.10
opencv-python>=4.8.0
Pillow>=10.0.0
pdf2image>=1.16.3
PyPDF2>=3.0.0
numpy>=1.24.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
llama-cpp-python>=0.2.0
spacy>=3.7.0
pyspellchecker>=0.7.2
python-dateutil>=2.8.2
sarvamai>=0.1.0
beautifulsoup4>=4.12.0
